add UT8_1 to TestSuite

+++ lab08/prj/ModulesMelnik/ModulesMelnik.h
@@ -1,6 +1,6 @@
 #ifndef MODULESMELNIK_H_INCLUDED
 #define MODULESMELNIK_H_INCLUDED
 
-float s_calculation(int x, int z);
+float s_calculation(float x, float z);
 
 #endif // MODULESMELNIK_H_INCLUDED
